+++ src/popup-container-provider/popup-container-provider.jsx
@@ -64,7 +64,7 @@ class PopupContainerProvider extends React.Component {
     };
 
     state = {
-        didRender: false
+        didRender: false // eslint-disable-line react/no-unused-state
     };
 
     getChildContext() {
@@ -100,7 +100,7 @@ class PopupContainerProvider extends React.Component {
      */
     handleContainerDidRender() {
         this.setState({
-            didRender: true
+            didRender: true // eslint-disable-line react/no-unused-state
         });
     }
 }