+++ CHANGELOG.md
@@ -1,3 +1,131 @@
+<a name="11.4.0"></a>
+# [11.4.0](https://github.com/alfa-laboratory/arui-feather/compare/v11.3.2...v11.4.0) (2017-12-27)
+
+
+### Bug Fixes
+
+* **button:** fix button alignment ([#433](https://github.com/alfa-laboratory/arui-feather/issues/433)) ([2ece9ff](https://github.com/alfa-laboratory/arui-feather/commit/2ece9ff))
+* **calendar:** fix Number.isFinite troubles in ie11 ([#436](https://github.com/alfa-laboratory/arui-feather/issues/436)) ([b6112a5](https://github.com/alfa-laboratory/arui-feather/commit/b6112a5))
+* **calendar:** update disabled days styles ([#434](https://github.com/alfa-laboratory/arui-feather/issues/434)) ([e2c8f4c](https://github.com/alfa-laboratory/arui-feather/commit/e2c8f4c))
+* **plate:** fix colors in theme-on-white ([#438](https://github.com/alfa-laboratory/arui-feather/issues/438)) ([3984ddc](https://github.com/alfa-laboratory/arui-feather/commit/3984ddc))
+
+
+### Features
+
+* **button:** remove view other prop ([#431](https://github.com/alfa-laboratory/arui-feather/issues/431)) ([f480b92](https://github.com/alfa-laboratory/arui-feather/commit/f480b92))
+* **calendar-input:** add max-min attribute for native calendar ([#429](https://github.com/alfa-laboratory/arui-feather/issues/429)) ([24a62b5](https://github.com/alfa-laboratory/arui-feather/commit/24a62b5))
+* **textarea:** add ability to control row amount ([#396](https://github.com/alfa-laboratory/arui-feather/issues/396)) ([1096059](https://github.com/alfa-laboratory/arui-feather/commit/1096059))
+
+
+
+<a name="11.3.2"></a>
+## [11.3.2](https://github.com/alfa-laboratory/arui-feather/compare/v11.3.1...v11.3.2) (2017-12-26)
+
+
+### Bug Fixes
+
+* enhance type and size props validation in checkbox & radio ([#402](https://github.com/alfa-laboratory/arui-feather/issues/402)) ([31ee0e4](https://github.com/alfa-laboratory/arui-feather/commit/31ee0e4))
+* **button:** fix icon alignment ([#430](https://github.com/alfa-laboratory/arui-feather/issues/430)) ([dc45392](https://github.com/alfa-laboratory/arui-feather/commit/dc45392))
+* **package:** update libphonenumber-js to the latest ([#413](https://github.com/alfa-laboratory/arui-feather/issues/413)) ([48f1a9a](https://github.com/alfa-laboratory/arui-feather/commit/48f1a9a)), closes [#412](https://github.com/alfa-laboratory/arui-feather/issues/412)
+
+
+
+<a name="11.3.1"></a>
+## [11.3.1](https://github.com/alfa-laboratory/arui-feather/compare/v11.3.0...v11.3.1) (2017-12-14)
+
+
+### Bug Fixes
+
+* **intl-phone-input:** optimizing intl-phone-input component ([#403](https://github.com/alfa-laboratory/arui-feather/issues/403)) ([0bd6c05](https://github.com/alfa-laboratory/arui-feather/commit/0bd6c05))
+
+
+
+<a name="11.3.0"></a>
+# [11.3.0](https://github.com/alfa-laboratory/arui-feather/compare/v11.2.1...v11.3.0) (2017-12-13)
+
+
+### Bug Fixes
+
+* **link:** fix security bug which work via target='_blank' ([#404](https://github.com/alfa-laboratory/arui-feather/issues/404)) ([badbd83](https://github.com/alfa-laboratory/arui-feather/commit/badbd83))
+
+
+### Features
+
+* **popup:** add gradient on top while scrolling ([#393](https://github.com/alfa-laboratory/arui-feather/issues/393)) ([19988b7](https://github.com/alfa-laboratory/arui-feather/commit/19988b7))
+
+
+
+<a name="11.2.1"></a>
+## [11.2.1](https://github.com/alfa-laboratory/arui-feather/compare/v11.2.0...v11.2.1) (2017-12-05)
+
+
+### Bug Fixes
+
+* **render-in-container:** use less loud deprecation warnings ([#395](https://github.com/alfa-laboratory/arui-feather/issues/395)) ([53d9ae0](https://github.com/alfa-laboratory/arui-feather/commit/53d9ae0))
+
+
+
+<a name="11.2.0"></a>
+# [11.2.0](https://github.com/alfa-laboratory/arui-feather/compare/v11.1.0...v11.2.0) (2017-12-05)
+
+
+### Bug Fixes
+
+* **flag-icon:** change root style to inline-flex ([#381](https://github.com/alfa-laboratory/arui-feather/issues/381)) ([64a114f](https://github.com/alfa-laboratory/arui-feather/commit/64a114f))
+* **intl-phone-input:** fix popup offset values ([#383](https://github.com/alfa-laboratory/arui-feather/issues/383)) ([a832794](https://github.com/alfa-laboratory/arui-feather/commit/a832794))
+* **intl-phone-input:** use input as root node ([#380](https://github.com/alfa-laboratory/arui-feather/issues/380)) ([61f8a69](https://github.com/alfa-laboratory/arui-feather/commit/61f8a69))
+* **menu:** fix item highlighting on componentWillReceiveProps ([#382](https://github.com/alfa-laboratory/arui-feather/issues/382)) ([382ccd5](https://github.com/alfa-laboratory/arui-feather/commit/382ccd5))
+* **popup:** fix server-side rendering ([#391](https://github.com/alfa-laboratory/arui-feather/issues/391)) ([477476c](https://github.com/alfa-laboratory/arui-feather/commit/477476c))
+
+
+### Features
+
+* support react 15 ([#394](https://github.com/alfa-laboratory/arui-feather/issues/394)) ([8538dad](https://github.com/alfa-laboratory/arui-feather/commit/8538dad))
+
+
+
+<a name="11.1.0"></a>
+# [11.1.0](https://github.com/alfa-laboratory/arui-feather/compare/v11.0.0...v11.1.0) (2017-11-29)
+
+
+### Bug Fixes
+
+* **popup:** fix server-side rendering ([#379](https://github.com/alfa-laboratory/arui-feather/issues/379)) ([b5ddba7](https://github.com/alfa-laboratory/arui-feather/commit/b5ddba7))
+
+
+### Features
+
+* update & freeze deps ([#374](https://github.com/alfa-laboratory/arui-feather/issues/374)) ([21c74b1](https://github.com/alfa-laboratory/arui-feather/commit/21c74b1))
+
+
+
+<a name="11.0.0"></a>
+# [11.0.0](https://github.com/alfa-laboratory/arui-feather/compare/v10.7.0...v11.0.0) (2017-11-27)
+
+
+### Features
+
+* migrate to react v16 ([#349](https://github.com/alfa-laboratory/arui-feather/issues/349)) ([bfbc9ee](https://github.com/alfa-laboratory/arui-feather/commit/bfbc9ee))
+* use only onClickOutside prop for autoclosable behavior ([#354](https://github.com/alfa-laboratory/arui-feather/issues/354)) ([f7c86a2](https://github.com/alfa-laboratory/arui-feather/commit/f7c86a2))
+
+
+
+<a name="10.7.0"></a>
+# [10.7.0](https://github.com/alfa-laboratory/arui-feather/compare/v10.6.0...v10.7.0) (2017-11-26)
+
+
+### Bug Fixes
+
+* **checkbox:** fix alignment without text ([#367](https://github.com/alfa-laboratory/arui-feather/issues/367)) ([38a155c](https://github.com/alfa-laboratory/arui-feather/commit/38a155c))
+
+
+### Features
+
+* add intl-phone-input component ([#345](https://github.com/alfa-laboratory/arui-feather/issues/345)) ([362ada9](https://github.com/alfa-laboratory/arui-feather/commit/362ada9))
+* add sizes for radio-group ([#369](https://github.com/alfa-laboratory/arui-feather/issues/369)) ([2429fbb](https://github.com/alfa-laboratory/arui-feather/commit/2429fbb))
+
+
+
 <a name="10.6.0"></a>
 # [10.6.0](https://github.com/alfa-laboratory/arui-feather/compare/v10.5.1...v10.6.0) (2017-11-17)